requirements.txt
altgraph==0.17
certifi==2020.12.5
chardet==4.0.0
future==0.18.2
idna==2.10
pefile==2019.4.18
Pillow==8.1.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pywin32-ctypes==0.2.0
selenium==3.141.0
soupsieve==2.1
urllib3==1.26.3
